# ggml : use F16 instead of F32 in Q4_0, Q4_1, Q8_0