No need to keep the scaling factor in F32 - use F16 instead to get smaller models and improve inference speed

This change makes the model files smaller, the inference faster and allows to fit more GPU layers in VRAM

- [x] check AVX
- [x] implement CUDA
- [x] recalc PPL 

New stats:

| Model | Measure      | F16    | Q4_0   | Q4_1   | Q5_0   | Q5_1   | Q8_0   |
|------:|--------------|-------:|-------:|-------:|-------:|-------:|-------:|
|    7B | perplexity   | 5.9066 | 6.1565 | 6.0912 | 5.9862 | 5.9481 | 5.9070 |
|    7B | file size    |  13.0G |   3.5G |   3.9G |   4.3G |   4.7G |   6.7G |
|    7B | ms/tok @ 4th |    127 |     55 |     54 |     76 |     83 |     72 |
|    7B | ms/tok @ 8th |    122 |     43 |     45 |     52 |     56 |     67 |
|    7B | bits/weight  |   16.0 |    4.5 |    5.0 |    5.5 |    6.0 |    8.5 |
|   13B | perplexity   | 5.2543 | 5.3860 | 5.3608 | 5.2856 | 5.2706 | 5.2548 |
|   13B | file size    |  25.0G |   6.8G |   7.6G |   8.3G |   9.1G |    13G |
|   13B | ms/tok @ 4th |    - |     103 |    105 |    148 |    160 |    131 |
|   13B | ms/tok @ 8th |    - |     73 |     82 |     98 |    105 |    128 |
|   13B | bits/weight  |   16.0 |    4.5 |    5.0 |    5.5 |    6.0 |    8.5 |

Old stats:

| Model | Measure      | F16    | Q4_0   | Q4_1   | Q5_0   | Q5_1   | Q8_0   |
|------:|--------------|-------:|-------:|-------:|-------:|-------:|-------:|
|    7B | perplexity   | 5.9066 | 6.1565 | 6.0910 | 5.9862 | 5.9481 | 5.9069 |
|    7B | file size    |  13.0G |   4.0G |   4.8G |   4.4G |   4.8G |   7.1G |
|    7B | ms/tok @ 4th |    128 |     50 |     54 |     75 |     83 |     75 |
|    7B | ms/tok @ 8th |    123 |     44 |     52 |     53 |     58 |     72 |
|    7B | bits/weight  |   16.0 |    5.0 |    6.0 |    5.5 |    6.0 |    9.0 |
|   13B | perplexity   | 5.2543 | 5.3860 | 5.3607 | 5.2856 | 5.2706 | 5.2548 |
|   13B | file size    |  25.0G |   7.6G |   9.1G |   8.4G |   9.1G |    14G |
|   13B | ms/tok @ 4th |    - |     93 |    101 |    150 |    164 |    141 |
|   13B | ms/tok @ 8th |    - |     81 |     96 |     96 |    104 |    136 |
|   13B | bits/weight  |   16.0 |    5.0 |    6.0 |    5.5 |    6.0 |    9.0 |

<!--
copilot:all
-->
### <samp>🤖 Generated by Copilot at d627025</samp>

### Summary
:compression::floppy_disk::arrows_counterclockwise:

<!--
1.  :compression: - This emoji represents the improvement in compression ratio that results from using a smaller data type for the delta field.
2.  :floppy_disk: - This emoji represents the reduction in memory footprint that results from using a smaller data type for the delta field.
3.  :arrows_counterclockwise: - This emoji represents the conversion between `float` and `ggml_fp16_t` that is required to perform arithmetic operations on the delta field.
-->
Use 16-bit floats for quantization deltas in `ggml.c` to save memory and compress better.

> _We're sailing on the sea of bits, we need to save some space_
> _We'll change the delta field to `ggml_fp16_t`_
> _Heave away, me hearties, heave away with grace_
> _We'll convert and swap and compress at a faster rate_

### Walkthrough
*  Replace the `float` type with the `ggml_fp16_t` type for the `delta` field of the `block_q4_0`, `block_q4_1`, `block_q8_0`, and `block_q8_1` structs in `ggml.c` to reduce the memory footprint and improve the compression ratio of the quantized blocks ([link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL751-R762), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL783-R792), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2640-R2638), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2906-R2900))
*  Add macro calls to `GGML_FP32_TO_FP16` to convert the `delta` value from a 32-bit float to a 16-bit half-float before storing it in the `y[i].d` field of the `block_q4_0`, `block_q4_1`, `block_q8_0`, and `block_q8_1` structs in `ggml.c` for various quantization schemes and SIMD implementations ([link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL819-R819), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL859-R860), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL991-R991), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL1026-R1026), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL1061-R1061), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2664-R2662), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2940-R2934))
*  Add macro calls to `GGML_FP16_TO_FP32` to convert the `delta`, `min`, and `sum` values from 16-bit half-floats to 32-bit floats before using them in the dequantization and dot product computations in `ggml.c` for various quantization schemes and SIMD implementations ([link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL1312-R1312), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL1332-R1333), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL1408-R1408), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2122-R2124), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2140-R2142), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2153-R2154), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2177-R2178), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2219-R2220), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2237-R2238), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2336-R2337), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2366-R2367), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2390-R2392), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2408-R2410), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2422-R2429), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2462-R2463), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2538-R2545), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2746-R2745), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2789-R2793), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2815-R2811), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2834-R2829), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2876-R2878), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2899-R2893))
*  Swap the order of the multiplication operands in the expressions `d * sum` and `sum * d` in `ggml.c` to avoid a potential loss of precision when multiplying a 32-bit float by a 16-bit half-float ([link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL1160-R1160), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2270-R2271), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2288-R2289), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2564-R2563), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2707-R2705), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2969-R2963))
*  Replace the calls to `_mm256_broadcast_ss` and `_mm_broadcast_ss` with the calls to `_mm256_set1_ps` and `_mm_set1_ps` in `ggml.c` to avoid loading the values from memory and instead use an immediate operand, which may improve the performance and reduce the memory access ([link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2153-R2154), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2177-R2178), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2219-R2220), [link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2237-R2238))
*  Add curly braces around the for loop body in `ggml.c` to improve the readability and consistency of the code style ([link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL1672-R1674))
*  Remove an extra space character between the `m4b` and `s8b` variable declarations in `ggml.c` to fix a minor formatting issue and align the code style with the rest of the file ([link](https://github.com/ggerganov/llama.cpp/pull/1508/files?diff=unified&w=0#diff-6d9ce99fcb6f51ff76f59e479f6e6fc0bb62edef7442805d7a5bb15b23996b5dL2093-R2095))