@LostRuins 2023-05-18
Hi again, just wondering if it will be able to transparently handle f32->f16 scaling factors when loading older quantizations (i.e. during model load)?
 
If not, and it breaks all old models, can I request an increase in LLAMA_FILE_VERSION so this change can be detected and handled if it ends up being merged?