@j-f1 2023-05-18
I believe that means "4 threads" @Green-Sky