@ilyakurdyukov 2023-05-19
My thoughts on model format changes: it's better to keep the short and portable model format for distribution, but make a script to convert that format to a format that's better suited for performance on a particular architecture. Also allow it to be loaded and converted in memory. Because, for example, for the performance of vector code, it is better to fuse four Q4_0:

```c
typedef struct {
	float d[4];
	uint8_t qs[4][QK4_0 / 2];
} block4_q4_0;
```

Or even 8 at once. So 4-8 `d` can be read with the one vector load instruction. And each field will be perfectly aligned for use with aligned vector load instructions. I think it's impossible to create an ideal model format for the best performance for every architecture.