@KerfuffleV2 2023-05-19
It kind of seems like the level of sacrifice required to allow `mmap`ing models is getting bigger and bigger. Maybe that should be the special case, where there's a tool to convert the model to something that can be `mmap`ed directly for people that really care about making loading a little bit faster.

Without the limitation of models having to be `mmap`able, doing small conversions online during the load process would be easy. Personally, I'm perfectly happy to trade slightly slower model loading for lower memory usage/faster inference and more options.