@ggerganov 2023-05-19
> Hi again, just wondering if backward compatibility is possible (e.g. transparently handle f32->f16 scaling factors when loading older quantizations during model load)?
> 
> If not, and it breaks all old models, can I request an increase in LLAMA_FILE_VERSION so this change can be detected and handled if it ends up being merged?
> 
> Edit: also looked at the chart, seems like q4_0 actually gets a regression in speed with this change?

This change breaks `Q4_0`, `Q4_1` and `Q8_0` models

`LLAMA_FILE_VERSION` has been increased to 3 together with `GGML_QNT_VERSION` to 2

The regression in `Q4_0` is only when using 4 threads. My explanations is that at 4 threads we are not yet memory bound and the F16 - > F32 conversion has a small negative impact on the overall performance due to extra compute. With more threads, the computation is memory bound, so we win by reducing the amount of memory used by the weights

Regarding the rest performance concerns by @ilyakurdyukov and @KerfuffleV2 
Since the computation on the CPU is memory-bound, this change actually improves the performance.
Additionally, having smaller weights allows to load more of them in VRAM which brings additional improvement when using GPU

Alternative memory layouts are being explored in #1073 and #1256