@KerfuffleV2 2023-05-19
My concern wasn't really about performance specifically, but the restriction of being tied to a `mmap`able file format. There isn't necessarily a one-size-fits-all approach to the model format that will be ideal for every CPU/GPU like ilyakurdyukov mentioned.

If the model file must be `mmap`able then you basically have to try to pick some middle ground even if there are small fixups that could easily be done. It also makes changes that break model file compatibility a lot more frequent, since the format _must_ be directly evaluable exactly as it exists on disk.

Now that there's GPU support, it may be even more of an issue. From what I know, GPUs (generally) like 16bit floats, x86 CPUs at least (generally) like 32bit floats. You have to choose one or the other at the moment, it seems, even if it's not ideal.