@YellowRoseCx 2023-05-19
Hey Mr @ggerganov since you merged this, would you mind increasing the LLAMA_FILE_VERSION so this change can be detected and handled for backwards compatibility with other models?

Do you have any throughts on transparently handling f32->f16 scaling factors when loading older quantizations during model load?