@SlyEcho 2023-05-19
Loading older models now results in a crash instead of an error message, as expected.