@Green-Sky 2023-05-20
```
$ ./main -m /mnt/nvme/ml/llama/ggjt_3/65B/llama-65B-q4_0-ggjt3.bin -ngl 16 -b 1 -n 10 -t 8 -s 3 -p "Once" -c 128
```
... I know I cheated a little, batchsize of 1, so it fits into vram (idk looks like it allocates vram based on batch size), only context size of 128


```
WARNING: when using cuBLAS generation results are NOT guaranteed to be reproducible.
main: build = 580 (3f008ca)
main: seed  = 3
llama.cpp: loading model from /mnt/nvme/ml/llama/ggjt_3/65B/llama-65B-q4_0-ggjt3.bin
llama_model_load_internal: format     = ggjt v3 (latest)
llama_model_load_internal: n_vocab    = 32000
llama_model_load_internal: n_ctx      = 128
llama_model_load_internal: n_embd     = 8192
llama_model_load_internal: n_mult     = 256
llama_model_load_internal: n_head     = 64
llama_model_load_internal: n_layer    = 80
llama_model_load_internal: n_rot      = 128
llama_model_load_internal: ftype      = 2 (mostly Q4_0)
llama_model_load_internal: n_ff       = 22016
llama_model_load_internal: n_parts    = 1
llama_model_load_internal: model size = 65B
llama_model_load_internal: ggml ctx size =    0,18 MB
llama_model_load_internal: mem required  = 31661,46 MB (+ 5120,00 MB per state)
llama_model_load_internal: [cublas] offloading 16 layers to GPU
llama_model_load_internal: [cublas] total VRAM used: 6949 MB
.....................
llama_init_from_file: kv self size  =  320,00 MB

system_info: n_threads = 8 / 24 | AVX = 1 | AVX2 = 1 | AVX512 = 0 | AVX512_VBMI = 0 | AVX512_VNNI = 0 | FMA = 1 | NEON = 0 | ARM_FMA = 0 | F16C = 1 | FP16_VA = 0 | WASM_SIMD = 0 | BLAS = 1 | SSE3 = 1 | VSX = 0 | 
sampling: repeat_last_n = 64, repeat_penalty = 1,100000, presence_penalty = 0,000000, frequency_penalty = 0,000000, top_k = 40, tfs_z = 1,000000, top_p = 0,950000, typical_p = 1,000000, temp = 0,800000, mirostat = 0, mirostat_lr = 0,100000, mirostat_ent = 5,000000
generate: n_ctx = 128, n_batch = 1, n_predict = 10, n_keep = 0


 Once again we are happy to be able to offer a
llama_print_timings:        load time = 34773,73 ms
llama_print_timings:      sample time =    15,87 ms /    10 runs   (    1,59 ms per token)
llama_print_timings: prompt eval time =     0,00 ms /     1 tokens (    0,00 ms per token)
llama_print_timings:        eval time = 101378,22 ms /    11 runs   ( 9216,20 ms per token)
llama_print_timings:       total time = 119457,45 ms
```

so I can run 65B with 32gig ram and 8gig vram with 9.2sec/tok

might get a little faster + 1 more layer on gpu, when there is no desktop session running..

:tada: